Major change, Added about Page and changed file structure
