+++ package.json
@@ -8,8 +8,10 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^5.5.0",
     "react-lottie": "^1.2.4",
     "react-reader": "^2.0.9",
+    "react-router-dom": "^7.4.1",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
