+++ src/pages/home.jsx
@@ -109,7 +109,7 @@ const Home = () => {
 
   return (
     <div
-      className="bg-white border-gray-200 dark:bg-teal-950 dark:text-white"
+      className="bg-white border-gray-200 dark:bg-teal-950 dark:text-white m-0"
       // style={{ height: "100vh" }}
     >
       <div className="p-6 mx-6 rounded-xl flex justify-between bg-yellow-50 dark:bg-teal-900 ">
@@ -140,13 +140,11 @@ const Home = () => {
         />
       </div>
       <div className="m-6 flex flex-wrap justify-around pb-16">
-        {/* <button className="px-6 m-2 py-3 bg-yellow-600 dark:bg-teal-700 hover:bg-slate-800 hover:dark:bg-teal-800 rounded text-white">
-          choose model
-        </button> */}
+        {/* Model Selection Dropdown - Improved Styling */}
         <div className="relative">
           <button
             id="dropdownDefaultButton"
-            className="text-white px-6 m-2 py-3 bg-yellow-600 dark:bg-teal-700 hover:bg-slate-800 hover:dark:bg-teal-800 rounded  focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium  text-sm text-center inline-flex items-center "
+            className="text-white px-6 m-2 py-3 bg-teal-700 hover:bg-teal-800 rounded-lg font-medium text-sm text-center inline-flex items-center border border-teal-600"
             type="button"
             onClick={() => {
               ToggleDropDown(!dropdown);
@@ -172,12 +170,12 @@ const Home = () => {
           </button>
 
           <div
-            className={`z-10 absolute ${
+            className={`z-10 absolute mt-1 ${
               dropdown || "hidden"
-            } bg-white divide-y divide-gray-100 rounded-lg shadow w-44 dark:bg-gray-700`}
+            } bg-teal-900 rounded-lg shadow-lg w-60 border border-teal-800`}
           >
             <ul
-              className="py-2 text-sm text-gray-700 dark:text-gray-200"
+              className="py-1 text-sm text-white"
               aria-labelledby="dropdownDefaultButton"
             >
               {models.map((item, index) => {
@@ -188,9 +186,10 @@ const Home = () => {
                         console.log(index)
                         setmodelno(index);
                         ToggleDropDown((dropdown) => !dropdown);
-
                       }}
-                      className="block px-4 py-2 hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white"
+                      className={`block w-full text-left px-4 py-3 hover:bg-teal-800 ${
+                        modelno === index ? "bg-teal-800" : ""
+                      }`}
                     >
                       {item.split('/')[1]}
                     </button>
@@ -220,7 +219,7 @@ const Home = () => {
         <div className="relative max-w-96">
           <button
             id="dropdownDefaultButton"
-            className="text-white px-6 m-2 py-3 bg-yellow-600 dark:bg-teal-700 hover:bg-slate-800 hover:dark:bg-teal-800 rounded  focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium  text-sm text-center inline-flex items-center "
+            className="text-white px-6 m-2 py-3 bg-teal-700 hover:bg-teal-800 rounded-lg font-medium text-sm text-center inline-flex items-center border border-teal-600"
             type="button"
             onClick={() => {
               ToggleDropDown2(!dropdown2);
@@ -247,12 +246,12 @@ const Home = () => {
           </button>
 
           <div
-            className={`z-10 absolute ${
+            className={`z-10 absolute mt-1 ${
               dropdown2 || "hidden"
-            } bg-white divide-y divide-gray-100 rounded-lg shadow w-44 dark:bg-gray-700`}
+            } bg-teal-900 rounded-lg shadow-lg w-60 border border-teal-800`}
           >
             <ul
-              className="py-2 text-sm text-gray-700 dark:text-gray-200"
+              className="py-1 text-sm text-white"
               aria-labelledby="dropdownDefaultButton"
             >
               {epubs.map((item, index) => {
@@ -264,7 +263,9 @@ const Home = () => {
                         setbookno(index);
                         ToggleDropDown2((dropdown2) => !dropdown2);
                       }}
-                      className="block px-4 py-2 hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white"
+                      className={`block w-full text-left px-4 py-3 hover:bg-teal-800 ${
+                        bookno === index ? "bg-teal-800" : ""
+                      }`}
                     >
                       {item.replace("/static/media/", "").split('.')[0]}
                     </button>
@@ -282,7 +283,7 @@ const Home = () => {
           Clear all
         </button>}
       </div>
-      <div className="flex flex-wrap p-6 m-6">
+      <div className="flex flex-wrap p-6 m-0">
         {image
           .slice(0)
           .reverse()
